NOTE: With debugging enabled, this may crash with failed assertions! I don't know precisely where it's legal to have NULL pointers, so I'm asserting every passed pointer to be not NULL.  Added assertions, no effect if '--enable-debug' is not passed to configure.
Note also that there are potential memory leaks and data loss errors
in here -- realloc() needs to be tested for its return value, since it
may fail.  As it stands, though, we're directly assigning the return
value of realloc() to the item that we're trying to resize, and this
may make that item invalid (NULL).

What is the best way to handle out-of-memory conditions?  This entire
file does an fprintf() to stderr, and then abort().  Should it call
panic() instead?